Improve footstep playback speed
The pitch change in the default footsteps when running is a bit exaggerated. This attempts to address the issue by lowering the change in pitch for walkspeeds higher or lower than 16.

Additionally, this corrects the usage of the deprecated [Sound.Pitch](https://developer.roblox.com/en-us/api-reference/property/Sound/Pitch) property. I originally couldn't find the part of the code that deals with footstep playback speed because I was searching for ".PlaybackSpeed" instead of ".Pitch"